Initial, non-working, new ML regression tests
The start of a new framework to replace the old;
the old one is strongly tied to the build environment,
and is very fragile.  Many of the tests have no description
for what they are testing.  Some tests run regardless of
which you select.  Some tests implicitly run on specific cams,
regardless of whether you have a rom or not, and regardless
of which cam you are trying to test.

Key differences that I hope to support:
 - allow the same testing system to work for devs and build / CI
 - failing tests should indicate why (to more detail than "didn't match expected logs")
 - for dev builds, detects available roms and only tests those
 - results are available in a machine readable format