Make EOSState s global, improve pointer syntax consistency
Making s global is prep for using plugins to monitor instructions;
specifically for DryosDebugMsg logging, though it is more general than that.
The old way using tb_exec_cb doesn't work in Qemu 4.

I think this is appropriate as a global since we only initialise it
in one place, there is only one entity of this type, and we pass it to
the majority of functions.  This is a global entity and can be treated as such.

Since I was looking over a lot of code I tried to change the
weird pattern of "char * name" into "char *name", as that's what
the rest of Qemu code uses (and it's the standard).