run_tests: 60D MenuTest, tweaks to allow it
Turns out 50D (while it is unreliable on qemu 2) is consistent
on qemu 4: but, it needs some time to initialise.  Thus reverted
the unreliable_screen_caps assignment.

Use of PIPE for capturing qemu output locks up if the pipe fills.
We're currently not using it, so changed to DEVNULL for now.
Later on we will want this for checking output.

Needed slightly higher wait after keypress, before screen cap,
or we kept capturing screen animation frames.