run_tests/qemu: replace vnc with qemu monitor
VNC was unreliable and hard code to work with.
Qemu monitor can entirely replace it, removing a source
of problems and a dependency.

I still see rare cases of "stuck screen" on e.g. 700D and 5D3,
and it's still the same screen each time, so this problem wasn't
VNC.  Suggests a rare emulation problem?